+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Fixed
+
+ - Fixed infinite loop in cursor when closed during read
+
 ## v2.2.1 - 2016-09-18
 
 ### Added